requirements.txt
allure-pytest==2.13.3
allure-python-commons==2.13.3
attrs==23.2.0
certifi==2024.2.2
h11==0.14.0
idna==3.6
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.0
pluggy==1.4.0
PySocks==1.7.1
pytest==8.1.1
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
